+++ src/controllers/SettingsController.php
@@ -136,17 +136,12 @@ class SettingsController extends BaseSettingsController
             $settings = Campaign::$plugin->settings;
         }
 
-        $memoryLimit = $settings->memoryLimit ? SendoutHelper::memoryInBytes($settings->memoryLimit) : 0;
-
         return $this->renderTemplate('campaign/_settings/sendout', [
             'settings' => $settings,
             'config' => Craft::$app->getConfig()->getConfigFromFile('campaign'),
             'contactElementType' => ContactElement::class,
-            'system' => [
-                'memoryLimit' => ini_get('memory_limit'),
-                'memoryLimitExceeded' => $memoryLimit > SendoutHelper::memoryInBytes(ini_get('memory_limit')),
-                'timeLimit' => ini_get('max_execution_time'),
-            ],
+            'memoryLimit' => ini_get('memory_limit'),
+            'timeLimit' => ini_get('max_execution_time'),
         ]);
     }
 
@@ -273,8 +268,6 @@ class SettingsController extends BaseSettingsController
         $settings->defaultNotificationContactIds = $this->request->getBodyParam('defaultNotificationContactIds', $settings->defaultNotificationContactIds) ?: null;
         $settings->showSendoutTitleField = $this->request->getBodyParam('showSendoutTitleField', $settings->showSendoutTitleField) ?: false;
         $settings->maxBatchSize = $this->request->getBodyParam('maxBatchSize', $settings->maxBatchSize) ?: null;
-        $settings->memoryLimit = $this->request->getBodyParam('memoryLimit', $settings->memoryLimit) ?: null;
-        $settings->timeLimit = $this->request->getBodyParam('timeLimit', $settings->timeLimit) ?: null;
 
         // Save it
         if (!Craft::$app->getPlugins()->savePluginSettings(Campaign::$plugin, $settings->getAttributes())) {