+++ src/services/SendoutsService.php
@@ -7,6 +7,7 @@ namespace putyourlightson\campaign\services;
 
 use Craft;
 use craft\base\Component;
+use craft\helpers\App;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\Db;
 use craft\helpers\Queue;
@@ -94,17 +95,6 @@ class SendoutsService extends Component
         }
 
         return $this->getSendoutById($sendoutId);
-    }
-
-    /**
-     * Returns sendout send status by ID.
-     */
-    public function getSendoutSendStatusById(int $sendoutId): bool|string|null
-    {
-        return SendoutRecord::find()
-            ->select('sendStatus')
-            ->where(['id' => $sendoutId])
-            ->scalar();
     }
 
     /**
@@ -600,8 +590,7 @@ class SendoutsService extends Component
      */
     private function getPendingRecipientsStandard(SendoutElement $sendout): array
     {
-        // Call for max power
-        Campaign::$plugin->maxPowerLieutenant();
+        App::maxPowerCaptain();
 
         $baseCondition = [
             'mailingListId' => $sendout->mailingListIds,